+++ src/value/index.ts
@@ -1,3 +1,3 @@
+export * from './track-value';
 export * from './value-sync';
 export * from './value-tracker';
-export * from './tracked-value';
